servlet, ch05

+++ jweb/src/main/webapp/address_book/addrDelete.jsp
@@ -2,10 +2,16 @@
     pageEncoding="UTF-8"%>
 <jsp:useBean id="abDAO" class="com.dao.AddrBookDAO" scope="application" />
 <%
+	
 	String uname = request.getParameter("username");
+
+	if(session.getAttribute("userName") != null){
+		abDAO.delete(uname);	//dao의 delete() 호출
+		response.sendRedirect("./addrList.jsp");		
+	}else{
+		response.sendRedirect("addrForm.jsp");	//목록 페이지로 이동
+	}
 	
-	abDAO.delete(uname);	//dao의 delete() 호출
 	
-	response.sendRedirect("addrList.jsp");	//목록 페이지로 이동
 
 %>
